requirements.txt
click==8.1.3
Flask==2.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2==2.9.3
PyJWT==2.4.0
Werkzeug==2.1.2
zipp==3.8.0
